publish: Use flatpak build-update-repo when needed
If the remote repo has runtimes, then we want to use flatpak
build-update-repo rather then ostree summary -u. Flatpak adds metadata
to the summary file that takes a long time to collect. However, it will
reuse the metadata if it's in the existing summary file. If we use
ostree summary -u, it will be lost and will need to be recollected the
next time flatpak build-update-repo is run on the server.