Ensure sane permissions when committing the tree for flatpak runtimes
This will improve the generation of static deltas, as we have patched
OSTree not to include in the deltas for files whose permissions in the
source commit are non readable by regular users, as that would cause
a permissions related problem when trying to apply it to produce the
updated file when pulling from the server as an unprivileged user.

While this might not seem very likely in the flatpak world, as flatpak
uses bare-user OSTree repositories to pull data from the server as an
unprivileged user (which will be handed over to the system helper to
actually deploy the update), it does affect Endless since the OSTree
repository for the entire OS, which is created in bare mode, is shared
with /var/lib/flatpak (it's actually the same repo). This would cause
trouble when updating runtimes if there's a delta to update a file that
is unreadable in the source commit, as the temporary repository created
at ~/.local/share/flatpak/system-cache will be bare-user with its parent
set pointing to /var/lib/flatpak (which is bare). This means that if the
source file is found in the parent reapo with unreadable permissions,
then the delta will fail and the update won't succeed.

For that reason we have included the patch in OSTree mentioned above,
but we still want to get as many files in the delta as possible, hence
the chmod line added here: to make sure all the files included in the
runtime are committed world-readable, so that OSTree's static delta
generation code will include them in the deltas once again (and this
is fine for a flatpak runtime as world-readable permissions is what
flatpak expects anyway, and the reason why it uses bare-user repos).

https://phabricator.endlessm.com/T14159