ostree: Move dpkg database into /usr
Once the OS /var gets mounted over the deployment /var at runtime, the
dpkg database is no longer accessible. This makes both dpkg and apt
useless as they don't know the system state.

Move the dpkg database into /usr/share/dpkg so that it can be seen at
runtime. A symlink in the deployment and tmpfiles.d configuration are
added so that a the tools will see the data unmodified. The symlink in
the deployment is just to help when inspecting a checkout without
booting it.

This is the same concept as rpm-ostree, which moves the rpm database
from /var/lib/rpm to /usr/share/rpm.

https://phabricator.endlessm.com/T14382