squash funky permissions from Flatpak runtimes
Remove setuid, setgid and sticky bits, as well as setting ownership to uid/gid 0, to match the permissions permitted by ostree commit --canonical-permissions, because Flatpak now enforces that no other permissions are present when it is pulling. Without this, flatpak >= 0.8.7 or >= 0.9.6 is unable to pull/deploy the Endless OS runtime any longer.

Use chmod/chown to avoid dependency on new ostree for now, as --canonical-permissions was only added in 2017.4 which isn't included in EOS 3.1.

https://phabricator.endlessm.com/T17654